feat(wintertc): migrate TextEncoder and TextDecoder from boa_runtime
Moves `boa_runtime::text` into `boa_wintertc::encoding`, the name the TC55
skeleton reserved for the Encoding Standard APIs, and points `boa_runtime` at
it. `boa_runtime::text` stays as a re-export under its historical name, so the
public API is unchanged.

`bytemuck` moves to `boa_wintertc` along with the module (it was its only
user), as does the `test-case` dev-dependency.
