+++ core/wintertc/Cargo.toml
@@ -14,11 +14,13 @@ rust-version.workspace = true
 boa_engine.workspace = true
 boa_gc.workspace = true
 base64.workspace = true
+bytemuck.workspace = true
 comfy-table.workspace = true
 rustc-hash = { workspace = true, features = ["std"] }
 
 [dev-dependencies]
 indoc.workspace = true
+test-case.workspace = true
 textwrap.workspace = true
 futures-lite.workspace = true
